CP-966 move TemplatesSelectModalProps to separate file
