+++ packages/app/src/templates/components/templates-select-modal/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export * from "./templates-select-modal.component"
+export * from "./templates-select-modal.interface"
+export * from "./templates-select-modal-test-ids.enum"
