+++ packages/app/src/templates/components/templates-select-modal/templates-select-modal.component.tsx
@@ -14,14 +14,10 @@ import Table, {
   Row,
   RowSize,
 } from "App/__deprecated__/renderer/components/core/table/table.component"
-import { Template } from "App/templates/dto"
-import { TemplatesSelectModalTestIds } from "App/templates/components/templates-select-modal/templates-select-modal-test-ids.enum"
-export interface TemplatesSelectModalProps {
-  open: boolean
-  onClose: () => void
-  onSelect: (template: Template) => void
-  templates: Template[]
-}
+import {
+  TemplatesSelectModalTestIds,
+  TemplatesSelectModalProps,
+} from "App/templates/components/templates-select-modal"
 
 const messages = defineMessages({
   title: { id: "module.templates.attachTemplateModalTitle" },
