+++ packages/app/src/templates/components/templates-select-modal/templates-select-modal.interface.ts
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+import { Template } from "App/templates/dto"
+
+export interface TemplatesSelectModalProps {
+  open: boolean
+  onClose: () => void
+  onSelect: (template: Template) => void
+  templates: Template[]
+}
